split up integration tests
