+++ pkg/tests/configure_integration_test.go
@@ -1,0 +1,276 @@
+package tests
+
+import (
+	"context"
+	"fmt"
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tests/pulcheck"
+	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tests/tfcheck"
+	"github.com/stretchr/testify/require"
+)
+
+func TestConfigureGetRawConfigDoesNotPanic(t *testing.T) {
+	t.Parallel()
+	// Regression test for [pulumi/pulumi-terraform-bridge#2262]
+	getOkExists := func(d *schema.ResourceData, key string) (interface{}, bool) {
+		v := d.GetRawConfig().GetAttr(key)
+		if v.IsNull() {
+			return nil, false
+		}
+		return d.Get(key), true
+	}
+	resMap := map[string]*schema.Resource{
+		"prov_test": {
+			Schema: map[string]*schema.Schema{
+				"test": {
+					Type:     schema.TypeString,
+					Optional: true,
+				},
+			},
+		},
+	}
+
+	runConfigureTest := func(t *testing.T, configPresent bool) {
+		tfp := &schema.Provider{
+			ResourcesMap: resMap,
+			Schema: map[string]*schema.Schema{
+				"config": {
+					Type:     schema.TypeString,
+					Optional: true,
+				},
+			},
+			ConfigureContextFunc: func(ctx context.Context, rd *schema.ResourceData) (interface{}, diag.Diagnostics) {
+				_, ok := getOkExists(rd, "config")
+				require.Equal(t, configPresent, ok, "Unexpected config value")
+				return nil, nil
+			},
+		}
+		bridgedProvider := pulcheck.BridgedProvider(t, "prov", tfp)
+		configVal := "val"
+		if !configPresent {
+			configVal = "null"
+		}
+		program := fmt.Sprintf(`
+name: test
+runtime: yaml
+resources:
+  prov:
+    type: pulumi:providers:prov
+	defaultProvider: true
+	properties:
+	  config: %s
+  mainRes:
+    type: prov:index:Test
+	properties:
+	  test: "hello"
+outputs:
+  testOut: ${mainRes.test}
+`, configVal)
+		pt := pulcheck.PulCheck(t, bridgedProvider, program)
+		pt.Up(t)
+	}
+
+	t.Run("config exists", func(t *testing.T) {
+		runConfigureTest(t, true)
+	})
+
+	t.Run("config does not exist", func(t *testing.T) {
+		runConfigureTest(t, false)
+	})
+}
+
+// TODO[pulumi/pulumi-terraform-bridge#2274]: Move to actual cross-test suite once the plumbing is done
+func TestConfigureCrossTest(t *testing.T) {
+	t.Parallel()
+	resMap := map[string]*schema.Resource{
+		"prov_test": {
+			Schema: map[string]*schema.Schema{
+				"test": {
+					Type:     schema.TypeString,
+					Optional: true,
+				},
+			},
+		},
+	}
+
+	runTest := func(t *testing.T, sch map[string]*schema.Schema, pulumiProgram, tfProgram string) {
+		var tfRd *schema.ResourceData
+		var puRd *schema.ResourceData
+		_ = puRd // ignore unused warning
+		tfp := &schema.Provider{
+			ResourcesMap: resMap,
+			Schema:       sch,
+			ConfigureContextFunc: func(ctx context.Context, rd *schema.ResourceData) (interface{}, diag.Diagnostics) {
+				if tfRd == nil {
+					tfRd = rd
+				} else {
+					puRd = rd
+				}
+
+				return nil, nil
+			},
+		}
+
+		tfdriver := tfcheck.NewTfDriver(t, t.TempDir(), "prov", tfp)
+		tfdriver.Write(t, tfProgram)
+		_, err := tfdriver.Plan(t)
+		require.NoError(t, err)
+		require.NotNil(t, tfRd)
+		require.Nil(t, puRd)
+
+		bridgedProvider := pulcheck.BridgedProvider(t, "prov", tfp)
+
+		pt := pulcheck.PulCheck(t, bridgedProvider, pulumiProgram)
+		pt.Preview(t)
+		require.NotNil(t, puRd)
+		require.Equal(t, tfRd.GetRawConfig(), puRd.GetRawConfig())
+	}
+
+	t.Run("string attr", func(t *testing.T) {
+		runTest(t,
+			map[string]*schema.Schema{
+				"config": {
+					Type:     schema.TypeString,
+					Optional: true,
+				},
+			},
+			`
+name: test
+runtime: yaml
+resources:
+	prov:
+		type: pulumi:providers:prov
+		defaultProvider: true
+		properties:
+			config: val
+	mainRes:
+		type: prov:index:Test
+		properties:
+			test: "val"
+`,
+			`
+provider "prov" {
+	config = "val"
+}
+
+resource "prov_test" "test" {
+	test = "val"
+}`)
+	})
+
+	t.Run("object block", func(t *testing.T) {
+		runTest(t,
+			map[string]*schema.Schema{
+				"config": {
+					Type:     schema.TypeList,
+					Optional: true,
+					Elem: &schema.Resource{
+						Schema: map[string]*schema.Schema{
+							"prop": {
+								Type:     schema.TypeString,
+								Optional: true,
+							},
+						},
+					},
+					MaxItems: 1,
+				},
+			},
+			`
+name: test
+runtime: yaml
+resources:
+	prov:
+		type: pulumi:providers:prov
+		defaultProvider: true
+		properties:
+			config: {"prop": "val"}
+	mainRes:
+		type: prov:index:Test
+		properties:
+			test: "val"
+`,
+			`
+provider "prov" {
+	config {
+		prop = "val"
+	}
+}
+
+resource "prov_test" "test" {
+	test = "val"
+}`)
+	})
+
+	t.Run("list config", func(t *testing.T) {
+		runTest(t,
+			map[string]*schema.Schema{
+				"config": {
+					Type:     schema.TypeList,
+					Optional: true,
+					Elem: &schema.Schema{
+						Type: schema.TypeString,
+					},
+				},
+			},
+			`
+name: test
+runtime: yaml
+resources:
+	prov:
+		type: pulumi:providers:prov
+		defaultProvider: true
+		properties:
+			configs: ["val"]
+	mainRes:
+		type: prov:index:Test
+		properties:
+			test: "val"
+`,
+			`
+provider "prov" {
+	config = ["val"]
+}
+
+resource "prov_test" "test" {
+	test = "val"
+}`)
+	})
+
+	t.Run("set config", func(t *testing.T) {
+		runTest(t,
+			map[string]*schema.Schema{
+				"config": {
+					Type:     schema.TypeSet,
+					Optional: true,
+					Elem: &schema.Schema{
+						Type: schema.TypeString,
+					},
+				},
+			},
+			`
+name: test
+runtime: yaml
+resources:
+	prov:
+		type: pulumi:providers:prov
+		defaultProvider: true
+		properties:
+			configs: ["val"]
+	mainRes:
+		type: prov:index:Test
+		properties:
+			test: "val"
+`,
+			`
+provider "prov" {
+	config = ["val"]
+}
+
+resource "prov_test" "test" {
+	test = "val"
+}`)
+	})
+}
